+++ src/osd.h
@@ -90,15 +90,15 @@ extern "C" {
 void *osd_batch_init(uint n);
 void osd_publish_batch(void *batch);
 void osd_add_bool_fact(void *batch, char const *name, osd_tag *tags, int n_tags, bool value);
-void osd_add_int_fact(void *batch, char const *name, osd_tag *tags, int n_tags, int value);
-void osd_add_uint_fact(void *batch, char const *name, osd_tag *tags, int n_tags, uint value);
+void osd_add_int_fact(void *batch, char const *name, osd_tag *tags, int n_tags, long value);
+void osd_add_uint_fact(void *batch, char const *name, osd_tag *tags, int n_tags, ulong value);
 void osd_add_double_fact(void *batch, char const *name, osd_tag *tags, int n_tags, double value);
 void osd_add_str_fact(void *batch, char const *name, osd_tag *tags, int n_tags, char *value);
 
 // Publish individual facts
 void osd_publish_bool_fact(char const *name, osd_tag *tags, int n_tags, bool value);
-void osd_publish_int_fact(char const *name, osd_tag *tags, int n_tags, int value);
-void osd_publish_uint_fact(char const *name, osd_tag *tags, int n_tags, uint value);
+void osd_publish_int_fact(char const *name, osd_tag *tags, int n_tags, long value);
+void osd_publish_uint_fact(char const *name, osd_tag *tags, int n_tags, ulong value);
 void osd_publish_double_fact(char const *name, osd_tag *tags, int n_tags, double value);
 void osd_publish_str_fact(char const *name, osd_tag *tags, int n_tags, char *value);
 #ifdef __cplusplus
